chore(improve feedback): Improve how feedback is displayed to users and admins
All feedback from the application is not focused whenever it's generated. After successful completion of tasks, users and admins are redirected to their profile pages. Error messages are displayed in red and success messages are displayed in green. Before users and admins can complete changing details of an order, a prompt will ask for confirmation and only execute if the users and admins confirm.

[finishes #163717837]